--- a/Engine/DataFeeds/Enumerators/EnqueueableEnumerator.cs
+++ b/Engine/DataFeeds/Enumerators/EnqueueableEnumerator.cs
@@ -19,6 +19,7 @@ using System.Collections;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Threading;
+using System.Threading.Tasks;
 
 namespace QuantConnect.Lean.Engine.DataFeeds.Enumerators
 {
@@ -39,6 +40,11 @@ namespace QuantConnect.Lean.Engine.DataFeeds.Enumerators
         private readonly int _timeout;
         private readonly object _lock = new object();
         private readonly BlockingCollection<T> _blockingCollection;
+
+        private int _count;
+        private int _lowerThreshold;
+        private Action _produce;
+        private Task _producer;
 
         /// <summary>
         /// Gets the current number of items held in the internal queue
@@ -120,6 +126,8 @@ namespace QuantConnect.Lean.Engine.DataFeeds.Enumerators
         /// <exception cref="T:System.InvalidOperationException">The collection was modified after the enumerator was created. </exception><filterpriority>2</filterpriority>
         public bool MoveNext()
         {
+            TriggerProducer();
+
             T current;
             if (!_blockingCollection.TryTake(out current, _timeout))
             {
@@ -136,7 +144,6 @@ namespace QuantConnect.Lean.Engine.DataFeeds.Enumerators
 
                 return !_end;
             }
-
             _current = current;
 
             // even if we don't have data to return, we haven't technically
@@ -189,6 +196,42 @@ namespace QuantConnect.Lean.Engine.DataFeeds.Enumerators
                 Stop();
                 if (_blockingCollection != null) _blockingCollection.Dispose();
                 _disposed = true;
+            }
+        }
+
+        /// <summary>
+        /// Sets the production <see cref="Action"/> and the lower threshold production trigger.
+        /// Will also start the first producer <see cref="Task"/>
+        /// </summary>
+        /// <param name="produce">The <see cref="Action"/> used to produce more items</param>
+        /// <param name="lowerThreshold">The threshold used to determine
+        /// if a new producer <see cref="Task"/> has to start</param>
+        public void SetProducer(Action produce, int lowerThreshold)
+        {
+            _producer = Task.Run(produce);
+            _produce = produce;
+            _lowerThreshold = lowerThreshold;
+        }
+
+        /// <summary>
+        /// If the <see cref="_produce"/> action was set <see cref="SetProducer"/>,
+        /// this method will generate a new task to execute the <see cref="_produce"/> action
+        /// when items are less than <see cref="_lowerThreshold"/>, the enumerator is not finished
+        /// <see cref="HasFinished"/> and previous <see cref="_producer"/> already finished running.
+        /// </summary>
+        private void TriggerProducer()
+        {
+            if (_produce != null
+                && !HasFinished
+                && _producer.IsCompleted
+                && _lowerThreshold > _count--)
+            {
+                // we use local count for the outside if, for performance, and adjust here
+                _count = Count;
+                if (_lowerThreshold > _count)
+                {
+                    _producer = Task.Run(_produce);
+                }
             }
         }
     }