… Common/Util/FuncParallelRunnerWorkItem.cs
… Common/Util/IParallelRunnerWorkItem.cs
… Common/Util/ParallelRunner.cs
… Common/Util/ParallelRunnerWorker.cs