<!--- Provide a general summary of your changes in the Title above -->

#### Description
<!--- Describe your changes in detail -->
- Deleting `ParallelRunnerController` and `ParallelRunnerWorker`.
Replacing them for `Tasks`. The consumer, `EnqueueableEnumerator` will
spin up a new producer Task when he is running low on items.

### Local performance testing

[10 Symbols Minute Data]
`PR`
28.01 seconds at 165k data points per second. Processing total of 4,627,697 data points.
27.78 seconds at 167k data points per second. Processing total of 4,627,697 data points.
26.95 seconds at 172k data points per second. Processing total of 4,627,697 data points.
`MASTER`
31.72 seconds at 146k data points per second. Processing total of 4,627,697 data points.
30.11 seconds at 154k data points per second. Processing total of 4,627,697 data points.
30.97 seconds at 149k data points per second. Processing total of 4,627,697 data points.

[400 Symbols Daily Data]
`PR`
23.90 seconds at 44k data points per second. Processing total of 1,052,359 data points.
25.34 seconds at 42k data points per second. Processing total of 1,052,359 data points.
24.86 seconds at 42k data points per second. Processing total of 1,052,359 data points.
`MASTER`
24.27 seconds at 43k data points per second. Processing total of 1,052,359 data points.
24.25 seconds at 43k data points per second. Processing total of 1,052,359 data points.
25.65 seconds at 41k data points per second. Processing total of 1,052,359 data points.

[HistoryRequestBenchmark]
`PR`
16.76 seconds at 54k data points per second. Processing total of 904,955 data points.
18.97 seconds at 48k data points per second. Processing total of 904,955 data points.
19.17 seconds at 47k data points per second. Processing total of 904,955 data points.
`MASTER`
22.00 seconds at 41k data points per second. Processing total of 904,955 data points.
20.86 seconds at 43k data points per second. Processing total of 904,955 data points.
20.78 seconds at 44k data points per second. Processing total of 904,955 data points.

**The PR provides a small speed improvement in some benchmarks**

[10 Symbols Minute Data]
`PR`
![pr_10_minute](https://user-images.githubusercontent.com/18473240/53752819-69a1cb80-3e8e-11e9-960b-0691634f354e.PNG)
`MASTER`
![master_10_minute](https://user-images.githubusercontent.com/18473240/53752820-69a1cb80-3e8e-11e9-8efa-b7061fcda037.PNG)

**The PR provides a 10% system CPU usage reduction**

#### Related Issue
<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps to reproduce -->
<!--- Please link to the issue here: -->
Closes https://github.com/QuantConnect/Lean/issues/2928
#### Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
The file system should not waste CPU time
#### Requires Documentation Change
<!--- Please indicate if these changes will require updates to documentation, and if so, specify what changes are required -->
N/A
#### How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->
Unit and regression tests
#### Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Non-functional change (xml comments/documentation/etc)

#### Checklist:
<!--- The following is a checklist of items that MUST be completed before a PR is accepted -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project.
- [x] I have read the **CONTRIBUTING** [document](https://github.com/QuantConnect/Lean/blob/master/CONTRIBUTING.md).
- [ ] I have added tests to cover my changes. <!--- If not applicable, please explain why -->
- [x] All new and existing tests passed.
- [x] My branch follows the naming convention `bug-<issue#>-<description>` or `feature-<issue#>-<description>`

<!--- Template inspired by https://www.talater.com/open-source-templates/#/page/99 -->